Move files within maplibre-winit directory
